+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.1.1 (2025-03-14)
+* Added layout handler
+* Fixed missing interface components
+
 ## v0.1.0 (2025-03-14)
 * Built initial Page structure
 * Added Harvest Transformer